requirements.txt
# Core
playwright>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
PyYAML>=6.0
httpx>=0.24.0

# LLM (Gemini 권장, 1개만 설치)
google-generativeai>=0.3.0

# Logging
structlog>=23.1.0
python-json-logger>=2.0.0

# Utilities
requests>=2.31.0
rich>=13.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
